+++ src/icalendar/alarms.py
@@ -88,6 +88,14 @@ class AlarmTime:
         return self._alarm
 
     @property
+    def action(self) -> str:
+        """The action invoked when this alarm triggers.
+
+        This delegates to :attr:`Alarm.ACTION <icalendar.cal.alarm.Alarm.ACTION>`.
+        """
+        return self.alarm.ACTION
+
+    @property
     def parent(self) -> Parent | None:
         """The component that contains the alarm.
 
@@ -190,7 +198,7 @@ class Alarms:
         if component is not None:
             self.add_component(component)
 
-    def add_component(self, component: Alarm | Parent):
+    def add_component(self, component: Alarm | Parent) -> None:
         """Add a component.
 
         If this is an alarm, it is added.
@@ -235,7 +243,7 @@ class Alarms:
         """Set the start of the component.
 
         If you have only absolute alarms, this is not required.
-        If you have alarms relative to the start of a compoment, set the start here.
+        If you have alarms relative to the start of a component, set the start here.
         """
         self._start = dt
 
@@ -243,7 +251,7 @@ class Alarms:
         """Set the end of the component.
 
         If you have only absolute alarms, this is not required.
-        If you have alarms relative to the end of a compoment, set the end here.
+        If you have alarms relative to the end of a component, set the end here.
         """
         self._end = dt
 
@@ -311,7 +319,7 @@ class Alarms:
     def _repeat(self, first: datetime, alarm: Alarm) -> Generator[datetime]:
         """The times when the alarm is triggered relative to start."""
         yield first  # we trigger at the start
-        repeat = alarm.REPEAT
+        repeat = alarm.repeat
         duration = alarm.DURATION
         if repeat and duration:
             for i in range(1, repeat + 1):
@@ -347,7 +355,7 @@ class Alarms:
         ]
 
     def _get_end_alarm_times(self) -> list[AlarmTime]:
-        """Return a list of alarm times relative to the start of the component."""
+        """Return a list of alarm times relative to the end of the component."""
         if self._end is None and self._end_alarms:
             raise ComponentEndMissing(
                 "Use Alarms.set_end because at least one alarm is relative to the end "